Updated routine for analyzing results from force extension simulations. Fixed bug in initializing bonds for linear chains.

+++ results/00_programs/python/analysis.py
@@ -17,6 +17,8 @@ forceExtension_parmcsv = '01_raw_data/forceExtension/forceExtension.csv'
 bendingPARM_parmcsv = '01_raw_data/bendingPARM/bendingPARM.csv'
 # default image resolution
 default_dpi = 200
+# generic tolerance for curve fitting
+TOL=.0001
 
 
 ## METHODS
@@ -156,7 +158,7 @@ def parse_results(parms = None, dir = None, simfile = None, col = None, row = No
     return parms
 
 # method for getting scaling simulation data (N vs. R)
-def plot_scaling(parms, N_col = None, R_col = None, fit = False, Title = None, X_label = None, Y_label = None, data_label = None, dpi = None, logscale_x = False, logscale_y = False, x_min = None, y_min = None, x_max = None, y_max = None, saveas = None, error = False, color = None, fit_hook = False):
+def plot_scaling(parms, N_col = None, R_col = None, fit = False, Title = None, X_label = None, Y_label = None, data_label = None, dpi = None, logscale_x = False, logscale_y = False, x_min = None, y_min = None, x_max = None, y_max = None, saveas = None, error = False, color = None, plot_log5 = False, plot_slope = False):
     # make sure that the proper parameters were passed to the method
     if N_col is None:
         exit()
@@ -165,6 +167,8 @@ def plot_scaling(parms, N_col = None, R_col = None, fit = False, Title = None, X
     # establish image parameters based on parameters passed to method
     if dpi is None:
         dpi = default_dpi
+    # set up for figure
+    fig, ax1 = plt.subplots()
     for c in range(len(R_col)):
         # get values, average duplicates
         x = []
@@ -173,49 +177,181 @@ def plot_scaling(parms, N_col = None, R_col = None, fit = False, Title = None, X
         for u in parms[N_col].unique():
             x.append(u)
             y_mean = parms[parms[N_col] == u]
-            y.append(y_mean[R_col[c] + '_M1'].mean())
-            if error:
-                v.append(y_mean[R_col[c] + '_var'].mean() / 2)
+            y.append(abs(y_mean[R_col[c] + '_M2'].mean()))
+            v.append(y_mean[R_col[c] + '_var'].mean() / 2)
         # plot values
         if error:
             # plot with error bars
             plt.errorbar(x, y, yerr = v, label = data_label[c], fmt = 'o', color = color[c])
         else:
-            plt.plot(x, y, 'o', label = data_label[c], fillstyle = 'none', color = color[c])
+            # plot without error bars
+            ax1.plot(x, y, 'o', label = data_label[c], fillstyle = 'none', color = color[c])
         # fit data to power low, if specified
         if fit:
             # fit data to power law equation, determine parameters
             popt, pcov = curve_fit(power_fit, x, y)
             x_fit = [ ((max(x) - min(x)) / (100 - 1)) * i + min(x) for i in range(100)]
             y_fit = [ power_fit(i, *popt) for i in x_fit]
-            plt.plot(x_fit, y_fit, '--', label = data_label[c] + " Fit", color = color[c])
+            ax1.plot(x_fit, y_fit, '--', label = data_label[c] + " Fit", color = color[c])
             # plt.plot([], [], ' ', label="A = {:.3f}".format(popt[0]))
-        if fit_hook:
-            # fit data to power law equation, determine parameters
-            A = y[15] / x[15]
-            x_fit = [ ((max(x) - min(x)) / (100 - 1)) * i + min(x) for i in range(100)]
-            y_fit = [ power_fit_hook(i, A) for i in x_fit]
-            plt.plot(x_fit, y_fit, '--', label = "Hookean Regime", color = "black", alpha = 0.5)
-        # if fit_pincus:
-        #     # fit data to power law equation, determine parameters
-        #     A = y[25]
+        if plot_log5 or plot_slope:
+            # if plotting either the log5 curve or the slope of the log5 curve fit
+            # convert that dataset to a logscale
+            x_log = []
+            y_log = []
+            v_log = []
+            for i in range(len(x)):
+                if abs(y[i]) < TOL or math.isnan(y[i]):
+                    continue
+                x_log.append(math.log(x[i]) / math.log(10.))
+                y_log.append(math.log(abs(y[i])) / math.log(10.))
+                if v[i] > TOL:
+                    v_log.append(v[i])
+                else:
+                    v_log.append(TOL)
+            # generate initial parameters for log5 curve
+            lb = [min(y_log), max(y_log) - (TOL / 100), -100, -100, -40, min(x_log) - 2 * TOL]
+            ub = [max(y_log), max(y_log), 200, 100, 40, min(x_log) - TOL]
+            parameterBounds = (lb, ub)
+            init = [min(y_log), max(y_log), 4., 50., 0.01, min(x_log) - (3. * TOL / 2.)]
+            # fit data to power log5 equation
+            popt, pcov = curve_fit(f = log5_fit, xdata =  x_log, ydata = y_log, p0 = init, bounds = parameterBounds, maxfev = 100000000) # sigma = v_log,
+            x_fit = [ ((max(x_log) - min(x_log)) / (100 - 1)) * i + min(x_log) for i in range(100)]
+            y_fit = [ log5_fit(i, *popt) for i in x_fit]
+            # calculate the slope of the line
+            if plot_slope:
+                # calculate the of the line
+                h = (max(x_fit) - min(x_fit)) / len(x_fit) # distance between points
+                x_d1 = []
+                y_d1 = []
+                for i in range(len(y_fit) - 1):
+                    x_d1.append(math.pow(10, (x_fit[i+1] + x_fit[i]) / 2.))
+                    y_d1.append((y_fit[i+1] - y_fit[i]) / h)
+            # convert everthing from the log scale, back the linear scale; plot
+            if plot_log5:
+                for i in range(len(x_fit)):
+                    x_fit[i] = math.pow(10, x_fit[i])
+                    y_fit[i] = math.pow(10, y_fit[i])
+                ax1.plot(x_fit, y_fit, '--', label = "log 5 Curve", color = color[c], alpha = 0.8)
+            if plot_slope:
+                # plot slope already calculated
+                ax2 = ax1.twinx()
+                ax2.plot(x_d1, y_d1, '-', label = "First Derivative of log5", color = color[c], alpha = 0.8)
+                ax1.plot([], [], '-', label = "First Derivative of log5", color = color[c], alpha = 0.8)
+                ax2.set_ylabel("Change in " + Y_label)
+                ax2.set_ylim(0, 3.)
+    ax1.set_xlim(x_min, x_max)
+    ax1.set_ylim(y_min, y_max)
+    if logscale_x:
+        ax1.set_xscale('log')
+    if logscale_y:
+        ax1.set_yscale('log')
+    ax1.legend(loc = 'upper left')
+    if Title is not None:
+        ax1.set_title(Title)
+    if Y_label is not None:
+        ax1.set_ylabel(Y_label)
+    if X_label is not None:
+        ax1.set_xlabel(X_label)
+    if saveas is not None:
+        fig.savefig(saveas, dpi = dpi, bbox_inches='tight')
+    plt.show()
+    plt.close()
+
+# method for plotting data from force extension simulations
+def plot_force_extension(parms, X_col = None, Y_col = None, iso_col = None, isolabel = None, logscale_x = False, logscale_y = False, x_min = None, x_max = None, y_min = None, y_max = None, X_label = None, Y_label = None, Title = None, saveas = None, plot_log5 = False, plot_data = False, plot_slope = False, dpi = None):
+
+    if X_col is None or Y_col is None or iso_col is None:
+        print("plot_force_extension :: Must specift columns from data frame to plot!")
+        exit()
+
+    # establish image parameters based on parameters passed to method
+    if dpi is None:
+        dpi = default_dpi
+
+    # parse data from data frame, plot
+    iso_vals = parms[iso_col].unique()
+    for iso in iso_vals:
+        # for each unique isolated value
+        iso_df = parms[parms[iso_col] == iso]
+        x = iso_df[X_col].tolist()
+        y = iso_df[Y_col + '_M2'].tolist()
+        v = iso_df[Y_col + '_var'].tolist()
+        y = [abs(i) for i in y]
+        # fit the data to a log5 curve
+        if plot_log5 or plot_slope:
+            # convert data to log scale
+            x_log = []
+            y_log = []
+            v_log = []
+            for i in range(len(x)):
+                if abs(y[i]) < TOL or math.isnan(y[i]):
+                    continue
+                x_log.append(math.log(x[i]) / math.log(10.))
+                y_log.append(math.log(abs(y[i])) / math.log(10.))
+                if v[i] > TOL:
+                    v_log.append(v[i])
+                else:
+                    v_log.append(TOL)
+            # initialize paramaters for curve fitting
+            lb = [min(y_log), max(y_log) - (TOL / 100), -100, -100, -40, min(x_log) - 2 * TOL]
+            ub = [max(y_log), max(y_log), 200, 100, 40, min(x_log) - TOL]
+            parameterBounds = (lb, ub)
+            init = [min(y_log), max(y_log), 4., 50., 0.01, min(x_log) - (3. * TOL / 2.)]
+            # fit data to log5 equation
+            popt, pcov = curve_fit(f = log5_fit, xdata =  x_log, ydata = y_log, p0 = init, bounds = parameterBounds, maxfev = 100000000) # sigma = v_log,
+            x_fit = [ ((max(x_log) - min(x_log)) / (100 - 1)) * i + min(x_log) for i in range(100)]
+            y_fit = [ log5_fit(i, *popt) for i in x_fit]
+            # calculate the slope of the line
+            if plot_slope:
+                h = (max(x_fit) - min(x_fit)) / len(x_fit) # distance between points
+                x_d1 = []
+                y_d1 = []
+                for i in range(len(y_fit) - 1):
+                    x_d1.append(math.pow(10, (x_fit[i+1] + x_fit[i]) / 2.))
+                    y_d1.append((y_fit[i+1] - y_fit[i]) / h)
+            # convert the log5 curve back to linear scale
+            for i in range(len(x_fit)):
+                x_fit[i] = math.pow(10, x_fit[i])
+                y_fit[i] = math.pow(10, y_fit[i])
+        # plot either the slope or the simulation data
+        if plot_slope:
+            # plot the slope of the line fit to the simulation data
+            plt.plot(x_d1, y_d1, '-', label = isolabel.format(iso))
+        else:
+            # plot the simulation data (and the log5 fit, if called)
+            if plot_data and plot_log5:
+                line = plt.plot(x_fit, y_fit, '--')
+                plt.plot(x, y, 'o', label = isolabel.format(iso), fillstyle = 'none', color = line[-1].get_color())
+            elif plot_log5:
+                line = plt.plot(x_fit, y_fit, '--', label = isolabel.format(iso))
+            elif plot_data:
+                plt.plot(x, y, 'o', label = isolabel.format(iso), fillstyle = 'none')
+    # finish formatting the graph, once all of the plots have been added
     plt.xlim(x_min, x_max)
     plt.ylim(y_min, y_max)
     if logscale_x:
         plt.xscale('log')
     if logscale_y:
         plt.yscale('log')
-    plt.legend(loc = 'upper left')
+    if plot_slope:
+        plt.legend(loc = 'upper left')
+    else:
+        plt.legend(loc = 'lower right')
     if Title is not None:
         plt.title(Title)
     if Y_label is not None:
-        plt.ylabel(Y_label)
+        if plot_slope:
+            plt.ylabel("Change in " + Y_label)
+        else:
+            plt.ylabel(Y_label)
     if X_label is not None:
         plt.xlabel(X_label)
     if saveas is not None:
         plt.savefig(saveas, dpi = dpi, bbox_inches='tight')
     plt.show()
     plt.close()
+
 
 
 # model equation for power law fitting
@@ -229,6 +365,10 @@ def power_fit_hook(x, A):
 # model equation for exponential decay fitting
 def exp_decay_fit(x, A, B):
     return A * (math.e ** (- x / B))
+
+# model equation for logistic5 equation
+def log5_fit (x, A, B, C, D, E, F):
+    return A + ((B - A) / ((1. + (C / (x - F)) ** D) ** E))
 
 # method for plotting the bond bond correlation for bending potential simulation data against s, the bond distance
 def plot_bendingpot_bbc (df = None, k = None, max_s = 10, plot_fit = False, saveas = None, logscale_y = True, Title = None, Y_label = None, X_label = None, dpi = None):
@@ -316,7 +456,7 @@ def plot_bending_lp (df = None, plot_expectation_CSA = False, plot_expectation_C
     CS_expect = []
     CSA_expect = []
     for i in range(len(k)):
-        y.append(-1 / math.log(l1[i] / math.pi))
+        y.append(-1 / math.log(l1[i]))
     if plot_expectation_CA:
         n = [i for i in range(1,100)]
         n_CA = [i for i in n]
@@ -335,7 +475,7 @@ def plot_bending_lp (df = None, plot_expectation_CSA = False, plot_expectation_C
             for j in range(3):
                 val = kdf.iloc[0]['l' + str(j) + '_M1']
                 x.append(j)
-                y.append(val / math.pi)
+                y.append(val)
             popt, pcov = curve_fit(exp_decay_fit, x, y)
             l_acc.append(popt[1])
         plt.scatter(k, l_acc, s=20, label = "$\\ell_{p, ac}$")
@@ -371,6 +511,8 @@ scaling = ("scaling" in sys.argv)
 forceExtension = ("forceExtension" in sys.argv)
 # perform bending parameter analysis
 bendingPARM = ("bendingPARM" in sys.argv)
+# update the simulation results, even if the simulation results file already exists
+update = ("update" in sys.argv)
 
 ## SCRIPT
 # load defaults / settings from yaml
@@ -410,27 +552,53 @@ if scaling:
 
 
 if forceExtension:
-    # get simulation parameters
-    FE_parms = pd.read_csv(forceExtension_parmcsv)
-    # average simulation properties
-    FE_parms = parse_results(parms = FE_parms, dir = '01_raw_data/forceExtension/', simfile = 'RE2E.dat', col = 1, title = 'E2Ex', M1 = True, M2 = True, var = True, bootstrapping = True, tabsep = True)
-    # save results
-    if not os.path.exists("02_processed_data/forceExtension/"):
-        os.mkdir("02_processed_data/forceExtension/")
-    FE_parms.to_csv("02_processed_data/forceExtension/forceExtension.csv")
-    exit()
+    if update or (not os.path.exists("02_processed_data/forceExtension/forceExtension.csv")):
+        # get simulation parameters
+        FE_parms = pd.read_csv(forceExtension_parmcsv)
+        # average simulation properties
+        FE_parms = parse_results(parms = FE_parms, dir = '01_raw_data/forceExtension/', simfile = 'RE2E.dat', col = 1, title = 'E2Ex', M1 = True, M2 = True, var = True, bootstrapping = True, tabsep = True)
+        FE_parms = parse_results(parms = FE_parms, dir = '01_raw_data/forceExtension/', simfile = 'RE2E.dat', col = 4, title = 'E2Etot', M1 = True, M2 = True, var = True, bootstrapping = True, tabsep = True)
+        FE_parms = parse_results(parms = FE_parms, dir = '01_raw_data/forceExtension/', simfile = 'ROG.dat', col = 4, title = 'ROGtot', M1 = True, M2 = True, var = True, bootstrapping = False, tabsep = True)
+        # save results
+        if not os.path.exists("02_processed_data/forceExtension/"):
+            os.mkdir("02_processed_data/forceExtension/")
+        FE_parms.to_csv("02_processed_data/forceExtension/forceExtension.csv")
+    else:
+        FE_parms = pd.read_csv("02_processed_data/forceExtension/forceExtension.csv")
     # perform analysis for each unqiue set of 'modules'
-    for mod in FE_parms['mod'].unique():
-        # isolate the parameters corresponding to the module
-        mod_parm = FE_parms[FE_parms['mod'] == mod]
-        # establish filenames, etc.
-        save_name = '02_processed_data/forceExtension/forceExtension_' + mod
-        if mod == "FElinearChainReal":
-            mod = "Real, Linear Polymer Chains"
-        elif mod == "FElinearChainIdeal":
-            mod = "Ideal, Linear Polymer Chains"
-        # plot x_distance
-        plot_scaling(mod_parm, N_col = 'F', R_col = ['E2Ex'], logscale_x = True, logscale_y = True, x_min = .00008, x_max = 10., y_min = 0.01, y_max = 500, X_label = "External Force ($f_{x}$)", Y_label = "Distance", data_label = ["Chain Extension in X-direction"], Title = "Force Extension for " + mod + " (N = 100)", saveas = save_name + '.png', error = False, color = ['tab:purple'], fit_hook = True)
+    # loop through each topology, bending potential, bending potential strength, and chain length
+    for top in FE_parms['R'].unique():
+        if top == 0:
+            top_NAME = "CHAIN"
+            # continue
+        elif top == 1:
+            top_NAME = "RING"
+        else:
+            print(f"forceExtension :: unknown TOP code {top}")
+            exit()
+        for pot in FE_parms['pot'].unique():
+            for N in FE_parms['N'].unique():
+                # create plot with force extension of bending constants merged
+                top_df = FE_parms[(FE_parms['R'] == top) & (FE_parms['pot'] == pot) & (FE_parms['N'] == N)]
+                if top_df.empty:
+                    continue
+                save_name = f"02_processed_data/forceExtension/FE_{top_NAME}_{pot}_N{N}"
+                # # plot force extension data for all bending constants, as well as log5 fit
+                # plot_force_extension(top_df, X_col = 'F', Y_col = 'E2Ex', iso_col = 'K', isolabel = '$k_{{\\theta}}$ = {:2}', logscale_x = True, logscale_y = True, x_min = .00008, x_max = 10., y_min = 0.01, y_max = 500, X_label = "External Force ($f_{x}$)", Y_label = "Chain Extension (X-Direction)", Title = f"Force Extension Data for {top_NAME} with {pot} potential (N = {N})", saveas = save_name + '_data.png', plot_data = True, plot_log5 = True)
+                # # plot slope of log5 fit for all bending constants
+                # plot_force_extension(top_df, X_col = 'F', Y_col = 'E2Ex', iso_col = 'K', isolabel = '$k_{{\\theta}}$ = {:2}', logscale_x = True, logscale_y = False, x_min = .00008, x_max = 10., y_min = 0., X_label = "External Force ($f_{x}$)", Y_label = "Chain Extension (X-Direction)", Title = f"Force Extension Slope for {top_NAME} with {pot} potential (N = {N})", saveas = save_name + '_slope.png', plot_log5 = True, plot_slope = True)
+                # # plot slope of log5 fit for all bending constants
+                # plot_force_extension(top_df, X_col = 'F', Y_col = 'E2Ex', iso_col = 'K', isolabel = '$k_{{\\theta}}$ = {:2}', logscale_x = True, logscale_y = True, x_min = .00008, x_max = 10., y_min = 0.01, y_max = 500, X_label = "External Force ($f_{x}$)", Y_label = "Chain Extension (X-Direction)", Title = f"Force Extension Curve for {top_NAME} with {pot} potential (N = {N})", saveas = save_name + '_log5.png', plot_log5 = True)
+                # create unique force extension plots for each unique bending constant
+                for k in FE_parms['K'].unique():
+                    # establish file names
+                    save_name = f"02_processed_data/forceExtension/FE_{top_NAME}_{pot}_N{N}_k{k}"
+                    # isolate the parameters corresponding to the set
+                    isoK_df = FE_parms[(FE_parms['R'] == top) & (FE_parms['pot'] == pot) & (FE_parms['N'] == N) & (FE_parms['K'] == k)]
+                    if isoK_df.empty:
+                        print(save_name + " does not exist!")
+                        continue
+                    plot_scaling(isoK_df, N_col = 'F', R_col = ['E2Ex'], logscale_x = True, logscale_y = True, x_min = .00008, x_max = 10., y_min = 0.01, y_max = 500, X_label = "External Force ($f_{x}$)", Y_label = "Chain Extension (X-Direction)", data_label = ["Simulation Data"], Title = f"Force Extension for {top_NAME} with {pot} potential ($k_{{\\theta}}$ = {k}, N = {N})", saveas = save_name + 'M2.png', error = False, color = ['tab:purple'], plot_log5 = False, plot_slope = False)
 
 if bendingPARM:
     # load parameters, add property calculations from simulations
